+++ pages/_app.tsx
@@ -22,8 +22,9 @@ function MyApp({ Component, pageProps }: AppProps) {
       <MediaContextProvider>
         <AtomicProvider
           data={{
+            translateIconsV2: true,
             iconsUrl:
-              'https://cdn-icons.occ.com.mx/atomic-icons-1.7.0-beta.5.svg'
+              'https://cdn-icons.occ.com.mx/atomic-icons-1.7.0-beta.7.svg'
           }}
         >
           <Component key={isClient} {...pageProps} />